Initial pass at separating Tournament entity from business logic
