+++ aesopstables.py
@@ -1,7 +1,7 @@
 from aesops import app, db
 from pairing.match import Match
 from pairing.player import Player
-from pairing.tournament import Tournament
+from data_models.tournaments import Tournament
 from aesops.user import User
